AB#92287 move type definition
